Avoid duplication of utils functions
