+++ clinica/pipelines/pet_volume/pet_volume_utils.py
@@ -1,6 +1,14 @@
-def init_input_node(pet_nii):
-    import nibabel as nib
+import typing as ty
+from os import getcwd
+from os.path import basename, join
+from pathlib import Path, PurePath
 
+import nibabel as nib
+import numpy as np
+from nibabel.nifti1 import Nifti1Header
+
+
+def init_input_node(pet_nii):
     from clinica.utils.filemanip import get_subject_id
     from clinica.utils.stream import cprint
     from clinica.utils.ux import print_begin_image
@@ -23,40 +31,93 @@ def init_input_node(pet_nii):
     return pet_nii
 
 
-def create_binary_mask(tissues, threshold=0.3):
-    from os import getcwd
-    from os.path import basename, join
-
-    import nibabel as nib
-    import numpy as np
-
+def _check_non_empty_tissue_list(tissues: ty.List[PurePath]) -> None:
+    """Check that provided list is non-empty."""
     if len(tissues) == 0:
         raise RuntimeError(
             "The length of the list of tissues must be greater than zero."
         )
 
+
+def _load_tissues(
+    tissues: ty.List[PurePath],
+) -> ty.Tuple[np.ndarray, np.ndarray, Nifti1Header]:
+    """Aggregates the image data contained in the tissue images provided.
+
+    Parameters
+    ----------
+    tissues : List[PurePath]
+        List of tissue images to aggregate.
+
+    Returns
+    -------
+    data : np.ndarray
+        Aggregated data.
+
+    affine : np.ndarray
+        Affine of the first image, acting as the affine
+        of the aggregated image under the assumption that
+        all images have the same affine.
+
+    header : Nifti1Header
+        Header of the aggregated image.
+    """
+    _check_non_empty_tissue_list(tissues)
     img_0 = nib.load(tissues[0])
     shape = list(img_0.get_fdata(dtype="float32").shape)
-
     data = np.zeros(shape=shape)
-
     for image in tissues:
-        data = data + nib.load(image).get_fdata(dtype="float32")
+        data += nib.load(image).get_fdata(dtype="float32")
+    return data, img_0.affine, img_0.header
 
+
+def create_binary_mask(
+    tissues: ty.List[PurePath],
+    threshold: float = 0.3,
+) -> PurePath:
+    """Create a binary mask Nifti1Image from the list of tissues.
+
+    Tissue images are summed and the result is thresholded with the
+    provided `threshold` input.
+
+    Parameters
+    ----------
+    tissues : List[PurePath]
+        List of paths to tissue Nifti1Images. Must be non-empty.
+
+    threshold : float, optional
+        Threshold to apply when binarizing the Nifti1Image.
+        Default=0.3.
+
+    Returns
+    -------
+    out_mask : PurePath
+        Path to the binary mask Nifti1Image.
+    """
+    data, affine, header = _load_tissues(tissues)
     data = (data > threshold) * 1.0
     out_mask = join(getcwd(), basename(tissues[0]) + "_brainmask.nii")
-
-    mask = nib.Nifti1Image(data, img_0.affine, header=img_0.header)
+    mask = nib.Nifti1Image(data, affine, header=header)
     nib.save(mask, out_mask)
-    return out_mask
+    return Path(out_mask)
 
 
-def apply_binary_mask(image, binary_mask):
-    from os import getcwd
-    from os.path import basename, join
+def apply_binary_mask(image: PurePath, binary_mask: PurePath) -> PurePath:
+    """Apply the provided `binary_mask` to the provided `image`.
 
-    import nibabel as nib
+    Parameters
+    ----------
+    image : PurePath
+        Path to the Nifti1Image to apply the mask on.
 
+    binary_mask : PurePath
+        Path to the Nifti1Image containing the mask.
+
+    Returns
+    -------
+    masked_image_path : PurePath
+        Path to the masked Nifti1Image.
+    """
     original_image = nib.load(image)
     mask = nib.load(binary_mask)
 
@@ -67,61 +128,80 @@ def apply_binary_mask(image, binary_mask):
         data, original_image.affine, header=original_image.header
     )
     nib.save(masked_image, masked_image_path)
-    return masked_image_path
+    return Path(masked_image_path)
 
 
-def create_pvc_mask(tissues):
+def create_pvc_mask(tissues: ty.List) -> PurePath:
+    """Create a pvc mask from tissue list.
 
-    from os import getcwd
-    from os.path import join
+    Parameters
+    ----------
+    tissues : List
+        List of paths to tissue Nifti1Images. Must be non-empty.
 
-    import nibabel as nib
-    import numpy as np
-
-    if len(tissues) == 0:
-        raise RuntimeError(
-            "The length of the list of tissues must be greater than zero."
-        )
-
-    img_0 = nib.load(tissues[0])
-    shape = img_0.get_fdata(dtype="float32").shape
-    background = np.zeros(shape=shape)
-
+    Returns
+    -------
+    out_mask : PurePath
+        Path to the resulting mask Nifti1Image.
+    """
+    background, affine, header = _load_tissues(tissues)
+    shape = background.shape
     shape += tuple([len(tissues) + 1])
     data = np.empty(shape=shape, dtype=np.float64)
-
-    for i in range(len(tissues)):
-        image = nib.load(tissues[i])
+    for i, tissue in enumerate(tissues):
+        image = nib.load(tissue)
         data[..., i] = np.array(image.get_fdata(dtype="float32"))
-        background = background + image.get_fdata(dtype="float32")
-
     background = 1.0 - background
     data[..., len(tissues)] = np.array(background)
-
     out_mask = join(getcwd(), "pvc_mask.nii")
-    mask = nib.Nifti1Image(data, img_0.affine, header=img_0.header)
+    mask = nib.Nifti1Image(data, affine, header=header)
     nib.save(mask, out_mask)
-    return out_mask
+    return Path(out_mask)
 
 
-def pet_pvc_name(pet_image, pvc_method):
+def pet_pvc_name(pet_image: PurePath, pvc_method: str) -> str:
+    """Build the name for the PET PVC interface.
+
+    Parameters
+    ----------
+    pet_image : PurePath
+        Path to the PET scan file.
+
+    pvc_method : str
+        Name of the PVC method. This will be concatenated
+        with the `pet_image` filename.
+
+    Returns
+    -------
+    pet_pvc_path : str
+        Name for the PET PVC interface.
+    """
     from os.path import basename
 
-    pet_pvc_path = "pvc-" + pvc_method.lower() + "_" + basename(pet_image)
-    return pet_pvc_path
+    return "pvc-" + pvc_method.lower() + "_" + basename(pet_image)
 
 
-def normalize_to_reference(pet_image, region_mask):
-    from os import getcwd
-    from os.path import basename, join
+def normalize_to_reference(pet_image: PurePath, region_mask: PurePath) -> PurePath:
+    """Normalize the provided `pet_image` by dividing by the mean
+    value of the region defined by the provided `region_mask`.
 
-    import nibabel as nib
-    import numpy as np
+    Parameters
+    ----------
+    pet_image : PurePath
+        Path to the Nifti1Image which should be normalized.
 
+    region_mask : PurePath
+        Path to the mask to be used to define the region.
+
+    Returns
+    -------
+    suvr_pet_path : PurePath
+        Path to the normalized Nifti1Image.
+    """
     pet = nib.load(pet_image)
     ref = nib.load(region_mask)
 
-    region = np.multiply(pet.get_fdata(dtype="float32"), ref.get_fdata(dtype="float32"))
+    region = pet.get_fdata(dtype="float32") * ref.get_fdata(dtype="float32")
     region_mean = np.nanmean(np.where(region != 0, region, np.nan))
 
     data = pet.get_fdata(dtype="float32") / region_mean
@@ -131,23 +211,28 @@ def normalize_to_reference(pet_image, region_mask):
     suvr_pet = nib.Nifti1Image(data, pet.affine, header=pet.header)
     nib.save(suvr_pet, suvr_pet_path)
 
-    return suvr_pet_path
+    return Path(suvr_pet_path)
 
 
-def atlas_statistics(in_image, in_atlas_list):
+def atlas_statistics(in_image: str, in_atlas_list: ty.List) -> ty.List:
     """Generate regional measure from atlas_list in TSV files.
 
     For each atlas name provided it calculates for the input image the mean
     for each region in the atlas and saves it to a TSV file.
 
-    Args:
-        in_image: A Nifti image
-        in_atlas_list: List of names of atlas to be applied
+    Parameters
+    ----------
+    in_image : str
+        Path to the Nifti image.
 
-    Returns:
-        List of paths to TSV files
+    in_atlas_list : List
+        List of names of atlas to be applied.
+
+    Returns
+    -------
+    atlas_statistics : List
+        List of paths to TSV files.
     """
-    from os import getcwd
     from os.path import abspath, join
 
     from nipype.utils.filemanip import split_filename
@@ -167,7 +252,6 @@ def atlas_statistics(in_image, in_atlas_list):
                 statistics_on_atlas(in_image, atlas_class(), out_atlas_statistics)
                 atlas_statistics_list.append(out_atlas_statistics)
                 break
-
     return atlas_statistics_list
